Cleanup errors (#88)
* Disambiguate whether a revalidator recognized a request when checking for a need to revalidate (#87)

* feat(revalidator): add handled bool

Previously, the revalidators were very ambiguous in identifying if they had actually handled a
request for a revalidation check or not -- a "nil, nil" or "voucher result, nil" was interpreted as
not being handled. This is unfortunate, cause particularly a "voucher result, nil" is an indication
the request is handled by this check, and we should stop processing other revalidators that might
override the value with "nil, nil". We now add a boolean to disambiguate whether the revalidator
"recognized" this request --- if true, processing stops.

* style(comments): fix types

* feat(impl): cleanup errors

cleanup errors -- move more error codes to proper types and insure single error dispatched

* test(impl): add unit test for incomplete response